backend/requirements.txt
flask==3.0.2
flask-cors==4.0.0
python-dotenv==1.0.1
requests==2.31.0
pillow==10.2.0
python-multipart==0.0.9
werkzeug==3.0.1
kaggle==1.5.16
python-jose==3.3.0 